Refactor transaction validator
